TIP-842: Move attribute option searchable repository to Akeneo\Pim\Structure (enrich bundle)
