+++ src/Pim/Bundle/EnrichBundle/Resources/config/repositories.yml
@@ -3,7 +3,7 @@ parameters:
     pim_enrich.repository.family.search.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\FamilySearchableRepository
     pim_enrich.repository.family_variant.search.class:   Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\FamilyVariantSearchableRepository
     pim_enrich.repository.attribute.search.class:        Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AttributeSearchableRepository
-    pim_enrich.repository.attribute_option.search.class: Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AttributeOptionSearchableRepository
+    pim_enrich.repository.attribute_option.search.class: Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\InternalApi\AttributeOptionSearchableRepository
 
     ## Default UI repositories
     pim_enrich.repository.category.class:                Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\CategoryRepository
@@ -15,7 +15,6 @@ parameters:
     pim_enrich.repository.locale.class:                  Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\LocaleRepository
     pim_enrich.repository.channel.class:                 Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\ChannelRepository
     pim_enrich.repository.currency.class:                Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\CurrencyRepository
-    pim_enrich.repository.attribute_group.search.class:  Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\InternalApi\AttributeGroupSearchableRepository
     pim_enrich.repository.job_instance.class:            Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobInstanceRepository
     pim_enrich.repository.job_execution.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobExecutionRepository
     pim_enrich.repository.job_tracker.class:             Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobTrackerRepository
@@ -117,13 +116,6 @@ services:
         arguments:
             - '@doctrine.orm.entity_manager'
             - '%akeneo_batch.entity.job_execution.class%'
-
-    pim_enrich.repository.attribute_option.search:
-        class: '%pim_enrich.repository.attribute_option.search.class%'
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '%pim_catalog.entity.attribute_option.class%'
-            - '@pim_catalog.repository.attribute'
 
     pim_enrich.repository.client:
         class: '%pim_enrich.repository.client.class%'
